Evitando repeticao de recompensa

+++ src/pages/gifts/cardGift.tsx
@@ -21,6 +21,11 @@ function buyAGift(gift: string) {
     gifts.map((g: any) => {
         if (g.gift === gift) {
             if(xp! >= g.xp ) {
+                let MyGiftAlreadyExists = myGifts.some((gift: string) => gift === g.gift )
+                if(MyGiftAlreadyExists) {
+                    alert('Você já adquiriu essa recompensa, conclua para obte-la novamente')
+                    return
+                }
                 let newXp = Number(xp) - g.xp
                 localStorage.setItem('xp', String(newXp))
                 if(!myGifts || myGifts.length < 1 ){
@@ -32,7 +37,7 @@ function buyAGift(gift: string) {
                 let newMyGifts = g.gift 
                 myGifts.push(newMyGifts)
                 localStorage.setItem('MyGifts', JSON.stringify(myGifts))
-                alert('Parbéns você adquiriu essa recompensa!')
+                alert('Parabéns você adquiriu essa recompensa!')
                 location.reload()
                 return
             }
